feat: Scripts to generate and run BIR
The run-bir script has some hard-coded paths at the moment, but
we should be able to get more general, or at least give a
list of instructions for getting the right setup as a followup.

This adds `gather` to the wheel, so we can run from the wheel.
For example

    $ lake build
    $ bin/make-wheel
    $ cd /tmp
    $ virtualenv env
    $ source env/bin/activate
    $ pip install ~/workplace/sherloc/KLR/.wheel/dist/klr_lang-0.0.7-cp313-cp313-macosx_14_0_arm64.whl
    $ env/lib/python3.13/site-packages/bin/klr nki-to-bir ~/workplace/sherloc/KLR/interop/test/test_nki_isa_tensor_scalar.py kernel1
    {"version": 2,
     "functions":
     [{"name": "sg0000",
     ...