refactor: remove Object type and Tracer monad
This change removes the ability to place lean functions inside the
trace environment, and simplifies the trace monad types. Builtin
functions are now represented in the environment by their unique
names. When a builtin is encountered, we use the name to dispatch to
the correct lean function. This dispatching is fixed and cannot be
changed at runtime.