feat: implement memory views
This is the second part of the new allocation API. This change adds memory
views as a mechanism for declaring tensors. Included is also a change to
builtin terms which allows them to carry a "self" term. This change is useful
for implementing object-like methods such as the memory view function.