feat: add invariants to Core types
This changes adds invariants, in the form of proof terms, to several of the
Core types. Once the types are constructed, the invariants can be relied upon.
As a general rule, we do not use the invariants directly in code that we want
to prove things about, to avoid dependent pattern matching.  Rather, we use an
error monad and then prove a theorem that guarantees the computation will not
fail by using the invariants.