chore: remove the triangle notation
The triangle notation was introduced to get around
a limitation in Lean that couldn't prove termination
when mapM was used, even if it was structurally
obvious. I didn't like the notation solution because
it seemed brittle; I tried replacing the notation with
a def and an abbrev, but neither worked. I still don't
understand why it works, but I also don't understand why

    | .tuple es .store => return .tuple (<- es.attach.mapM fun ⟨ e, _ ⟩ => LValue e)

works but

    | .tuple es .store => return .tuple (<- es.attach.mapM fun e => LValue e.val)

does not. Since this is now required in only 2 spots (thanks
likely to improvements in Lean itself) I propose we just remove
the notation and do the trick until it's no longer needed.