feat: introduce NKI abstract syntax
This change introduces the abstract syntax of NKI.  This change is part of a
larger effort to break KLR up into smaller passes that can be more easily
compared with the upcoming C++ implementation. The first of these passes will
be a simplification pass that consumes the Python Core AST and produces the NKI
AST (contained in this patch).