feat: add support for assignment to subscript (x[s] = e)
This patch adds an assignment to subscript 'x[s] = e'.
This is an important example because, when x is a location to the memory and e is
nki.language.load, it must immediately store to the location.
Previously, when the RValue function is traversing e, it could have created a temporary
access (Tensor.load), so this could not be achieved.
This problem is addressed in this patch by making the `assign` function
not call `RValue` and making `assignTerm` function probe whether the RHS was `.store`.