feat: trace `x = f()` and `x binop= f()` with fn keyword args
This patch supports an alternative form of tracing `x = f()` and `x binop= f()` when this is necessary.

Specifically, this is for supporting lowering of `res_psum += nl.matmul(...)` NKI to KLR.

The Python tracer was first unfolding this to `res_psum = res_psum + nl.matmul(...)`, but this has two issues:

1. It has to expand `exprOp : BinOp -> Expr -> Expr -> Trace Term` when the second argument of exprOp is Core.Expr.call.
However, in this case, it is not clear how the output of `exprOp` can be represented in Term.
One possible solution is to temporarily store the output of `nl.matmul` into a temporary buffer
(in SBUF or PSUM or anywhere) and use the access as the second argument of `exprOp`,
but then the following issue appears.

2. I understood that `res_psum += nl.matmul(...)` is supposed to do an in-place update to `res_psum`
(but without a clear reason; I looked into the s3d3_mm doc but could not immediately find a flag for this.
But it is kind of weird if '+=' tries to store into a new memory region :/).
Storing the output `nl.matmul` into a temporary buffer does not match what NKI does.

To avoid this problem, this patch tries to achieve the following lowerin:

```
t += matmul(a,b)  # becomes matmul(a, b, dst = t, accum = True)
t = matmul(a,b)   # becomes matmul(a, b, dst = t, accum = False)
```

This patch detects this `y (binop)= f()` pattern from KLR.Python.Stmt and tries to evaluate
the alternative `f(..., dst = t(, accum = ...))` function call.
It only does this when the function is built-in or non-user-Python function.