chore: Have FromSexp throw on unknown fields
I've had several bugs with typos in names. This is especially
problematic when you rename a field in Lean, but forget to
update the name in text files. This change throws on unknown
names.

It also introduces a small Meta library that can be a home
for common patterns of metaprogramming we use, so we don't
need to remember dozens of library names.