fix: chained comparison should not short-circuit
In Python, a chained comparison does not short-circuit. For example, if you
write the expression:

  1 < 0 < f()

then the function f will be called. This is different from logical and/or which
does short-circuit. We also have the possibility of logical and/or on tensors,
which does not short-circuit. This change introduces two new expression forms
`conj` and `disj` which are used to represent Python short-circuit and/or. The
normal land and lor binary operator do not short-circuit. This gives us what we
need to lower logical operators, tensor operators, and chained comparison.