feat: implement LNC kernels
This change implement basic LNC kernels for grid with only one dimension.  When
a user requests an LNC grid of size N, then we will evaluate the kernel N
times.  For each evaluation we set the program_id to a value from zero to N.
This create N different kernel artifacts which we encode in a list.