fix: gather supports numpy.ndarray
numpy.ndarray was being rejected. It's tp_name is "numpy.ndarray" with a ".". Other tensor types we support do not have a "." in their tp_name. (The difference isn't affected by python version, I tested 3.9-3.13).  Maybe we shouldn't be checking type names, maybe we should just "duck type" and check if they have a shape and dtype? We can look into that in the future... this simple change unblocks numpy.