feat: add hierarchical debug info for profiling
This change introduce a new debug info format that is structured as a tree that
aligns with the lexical scope of the original program. There is also a
mechanism to set the filename for the debug output. If no file name is given,
then no debug information will be computed.