fix: more efficient CBOR encoding
The to- and from-CBOR functions are pure functions. This is convenient, but for
very large programs (> 1GB), the to-CBOR function is very inefficient. This
change introduces a second, more efficient CBOR encoding function that is
tail-recursive, and directly writes to a file handle rather than building an
intermediate array. There is a default implementation that uses the original
function, so implementers do not have to make any changes. However, the
instances for basic types (like List) and the generated instances have
specialized, efficient versions.